requirements.txt
# Embeddings & Vector Store
chromadb>=0.4.22
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.2

# RAG Framework
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.1

# Utilities
python-dotenv>=1.0.0

# Backend API (Phase 5)
fastapi>=0.109.0
uvicorn>=0.27.0

# Frontend (Phase 5)
streamlit>=1.30.0

# Development
pytest>=7.4.0
